Refactor2 (#196)
* Refactor
** Split parts of too large interfaces/validator/index -file to validations and log-actions files

* Update deps

* 3.3.10-alpha.3

---------
