+++ package.json
@@ -14,7 +14,7 @@
 		"url": "git@github.com:NatLibFi/melinda-rest-api-validator.git"
 	},
 	"license": "AGPL-3.0+",
-	"version": "3.3.9",
+	"version": "3.3.10-alpha.3",
 	"main": "./dist/index.js",
 	"engines": {
 		"node": ">=18"
@@ -34,7 +34,7 @@
 		"dev": "NODE_ENV=development cross-env nodemon"
 	},
 	"dependencies": {
-		"@babel/runtime": "^7.21.5",
+		"@babel/runtime": "^7.22.3",
 		"@natlibfi/marc-record": "7.2.2",
 		"@natlibfi/marc-record-merge": "7.0.0-alpha.1",
 		"@natlibfi/marc-record-serializers": "^10.0.4",
